requirements.txt
discord.py
requests
python-dotenv
validators
python-dateutil
python-Levenshtein
pytz